+++ server/run_server.jl
@@ -1,0 +1,3 @@
+using WflowServer
+
+WflowServer.main()