+++ docs/src/changelog.md
@@ -17,6 +17,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   lake storage. The variable `actevap` has also been added to the reservoir module.
 - The `set_states` function for model type `sbm` with local inertial routing for river and
   land component.
+- Inflow to reservoir and lake locations for local inertial routing: 1) with floodplain
+  routing, the floodplain discharge was not added to the inflow of these locations, 2) the
+  `to_river` variable from overland flow and lateral subsurface flow was not added to the
+  inflow of these locations.
 
 ### Changed
 - Stop exposing scalar variables through BMI. The `BMI.get_value_ptr` function was
@@ -29,6 +33,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Total water storage as an export variable for `SBM` concept. This is the total water stored
   per grid cell in millimeters. Excluded from this variable are the floodplain, lakes and
   reservoirs.
+- Checks to see if all states are covered in the .toml file. If not all states are covered,
+  an error is thrown. If there are more states specified than required, these states are
+  ignored (with a warning in the logging), and the simulation will continue.
+- Support different vertical hydraulic conductivity profiles for the `SBM` concept. See also
+  the following sections: [The SBM soil water accounting scheme](@ref) and [Subsurface flow
+  routing](@ref) for a short description.
 
 ## v0.7.3 - 2024-01-12
 
@@ -47,7 +57,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   terms of the quadratic equation (and solution) were fixed.
 - Use `kvfrac` for the computation of vertical saturated hydraulic conductivity at the
   bottom of the soil layer, since `kvfrac` is also used for the computation of vertical
-  unsaturated flow. 
+  unsaturated flow.
 
 ### Changed
 - For cyclic parameters different cyclic time inputs are supported (only one common cyclic